Pushed v1.24 code
- BPM is now slightly more accurate to CIA specifications, and also more accurate in vblank (VBL) timing mode using the true Amiga PAL vblank rate. NOTE: We're speaking tiny variations in BPM here...
- Vblank (VBL) timing mode: Fixed a bug with setting a speed of FF (255) on the first row. The tick duration would be about twice as long...
- Small code cleanup

+++ src/pt2_mouse.c
@@ -1048,7 +1048,7 @@ void sampleRepeatLengthDownButton(bool fast)
 
 void tempoUpButton(void)
 {
-	int16_t val;
+	int32_t val;
 
 	if (editor.timingMode == TEMPO_MODE_VBLANK)
 		return;
@@ -1069,7 +1069,7 @@ void tempoUpButton(void)
 
 void tempoDownButton(void)
 {
-	int16_t val;
+	int32_t val;
 
 	if (editor.timingMode == TEMPO_MODE_VBLANK)
 		return;
